requirements.txt
accelerate==0.21.0
gradio==3.39.0
matplotlib==3.4.3
numpy==1.25.2
onnxruntime==1.15.1
opencv_python==4.5.4.58
pandas==1.3.4
Pillow==10.0.0
pymilvus==2.2.14
PyYAML==6.0.1
redis==4.6.0
Requests==2.31.0
scikit_learn==1.0.2
torch==1.13.0
tqdm==4.62.3
transformers==4.31.0
